Imt_Provioder_Service_CRUD_Completed

+++ SharedKernel/src/SharedKernel.Main/Domain/IMT/ProviderService.cs
@@ -2,9 +2,9 @@
 
 public partial class ProviderService
 {
-    public int Id { get; set; }
+    public ulong Id { get; set; }
 
-    public int ImtProviderId { get; set; }
+    public ulong ImtProviderId { get; set; }
 
     /// <summary>
     /// BankAccount, WalletTransfer etc
